+++ src/pages/DaoDashboard/Info/InfoProposals.tsx
